Added injection for list of objects
