+++ build.gradle.kts
@@ -13,12 +13,13 @@ plugins {
 
     id(Plugins.update_dependencies) version PluginVers.update_dependencies
     id(Plugins.detekt) version PluginVers.detekt
+    id(Plugins.jacoco)
 
     `maven-publish`
 }
 
 group = "com.github.mynameisscr"
-version = "0.0.1"
+version = "v0.0.2"
 
 repositories {
     mavenCentral()
@@ -29,6 +30,7 @@ dependencies {
 
     testImplementation(Libs.kotlin_test)
     testImplementation(Libs.kotest_assertions_core_jvm)
+    testImplementation(Libs.mockk)
 }
 
 java {
@@ -123,5 +125,21 @@ tasks {
             showStandardStreams = true
             exceptionFormat = TestExceptionFormat.FULL
         }
+
+        finalizedBy(this@tasks.jacocoTestReport)
+    }
+
+    jacocoTestReport {
+        dependsOn(withType<Test>())
+
+        classDirectories.setFrom(
+            files(classDirectories.files.map {
+                fileTree(it) {
+                    exclude(
+                        "**/uuid/**",
+                    )
+                }
+            })
+        )
     }
 }
