+++ buildSrc/src/main/kotlin/Plugins.kt
@@ -11,4 +11,6 @@ object Plugins {
 
     const val detekt = "io.gitlab.arturbosch.detekt"
     const val detekt_formatting = "io.gitlab.arturbosch.detekt:detekt-formatting"
+
+    const val jacoco = "jacoco"
 }
