+++ src/main/kotlin/com/github/mynameisscr/kinja/exception/CandidateIsAlreadyExistsException.kt
@@ -1,0 +1,6 @@
+package com.github.mynameisscr.kinja.exception
+
+import kotlin.reflect.KType
+
+class CandidateIsAlreadyExistsException(name: String, returnType: KType) :
+    RuntimeException("Candidate with type '$returnType' and name '$name' is already exists")
