+++ src/main/kotlin/com/github/mynameisscr/kinja/exception/CandidateWithoutNameException.kt
@@ -1,0 +1,3 @@
+package com.github.mynameisscr.kinja.exception
+
+class CandidateWithoutNameException : RuntimeException("Candidate to inject has to have name")
