+++ src/main/kotlin/com/github/mynameisscr/kinja/exception/CircularDependencyException.kt
@@ -1,0 +1,3 @@
+package com.github.mynameisscr.kinja.exception
+
+class CircularDependencyException(chain: String) : RuntimeException("Detected circular dependency: $chain")
