+++ src/main/kotlin/com/github/mynameisscr/kinja/exception/ContextIsAlreadyInitializedException.kt
@@ -1,0 +1,3 @@
+package com.github.mynameisscr.kinja.exception
+
+class ContextIsAlreadyInitializedException : RuntimeException("Context is already initialized")
