+++ src/main/kotlin/com/github/mynameisscr/kinja/exception/ContextIsNotInitializedException.kt
@@ -1,0 +1,3 @@
+package com.github.mynameisscr.kinja.exception
+
+class ContextIsNotInitializedException : RuntimeException("Context isn't initialized")
