+++ src/main/kotlin/com/github/mynameisscr/kinja/exception/MultipleObjectsWithTypeException.kt
@@ -1,0 +1,6 @@
+package com.github.mynameisscr.kinja.exception
+
+import kotlin.reflect.KClass
+
+class MultipleObjectsWithTypeException(clazz: KClass<*>) :
+    RuntimeException("Found multiple objects by class ${clazz.qualifiedName}")
