+++ src/main/kotlin/com/github/mynameisscr/kinja/holder/Holder.kt
@@ -1,0 +1,6 @@
+package com.github.mynameisscr.kinja.holder
+
+interface Holder<T> {
+
+    fun get(): T
+}
