+++ src/main/kotlin/com/github/mynameisscr/kinja/holder/SingleHolder.kt
@@ -1,0 +1,6 @@
+package com.github.mynameisscr.kinja.holder
+
+class SingleHolder<T>(private val obj: T) : Holder<T> {
+
+    override fun get(): T = obj
+}
