+++ src/main/kotlin/com/github/mynameisscr/kinja/inject/InjectCandidate.kt
@@ -1,0 +1,15 @@
+package com.github.mynameisscr.kinja.inject
+
+import java.util.UUID
+import kotlin.reflect.KFunction
+
+data class InjectCandidate(
+    val id: UUID,
+    val name: String,
+    val func: KFunction<Any>,
+    val scope: InjectScope,
+) {
+
+    val params = func.parameters
+    val returnType = func.returnType
+}
