+++ src/main/kotlin/com/github/mynameisscr/kinja/inject/InjectScope.kt
@@ -1,0 +1,6 @@
+package com.github.mynameisscr.kinja.inject
+
+enum class InjectScope {
+    SINGLE,
+    FACTORY,
+}
