+++ src/main/kotlin/com/github/mynameisscr/kinja/uuid/UUIDGenerator.kt
@@ -1,0 +1,8 @@
+package com.github.mynameisscr.kinja.uuid
+
+import java.util.UUID
+
+interface UUIDGenerator {
+
+    fun generate(): UUID
+}
