+++ src/main/kotlin/com/github/mynameisscr/kinja/uuid/UUIDGeneratorImpl.kt
@@ -1,0 +1,8 @@
+package com.github.mynameisscr.kinja.uuid
+
+import java.util.UUID
+
+class UUIDGeneratorImpl : UUIDGenerator {
+
+    override fun generate(): UUID = UUID.randomUUID()
+}
