--- src/main/kotlin/ru/scratty/kinja/exception/CandidateIsAlreadyExistsException.kt
@@ -1,6 +1,0 @@
-package ru.scratty.kinja.exception
-
-import kotlin.reflect.KType
-
-class CandidateIsAlreadyExistsException(name: String, returnType: KType) :
-    RuntimeException("Candidate with type '$returnType' and name '$name' is already exists")
