--- src/main/kotlin/ru/scratty/kinja/exception/CandidateWithoutNameException.kt
@@ -1,3 +1,0 @@
-package ru.scratty.kinja.exception
-
-class CandidateWithoutNameException : RuntimeException("Candidate to inject has to have name")
