--- src/main/kotlin/ru/scratty/kinja/exception/CircularDependencyException.kt
@@ -1,3 +1,0 @@
-package ru.scratty.kinja.exception
-
-class CircularDependencyException(chain: String) : RuntimeException("Detected circular dependency: $chain")
