--- src/main/kotlin/ru/scratty/kinja/exception/ContextIsAlreadyInitializedException.kt
@@ -1,3 +1,0 @@
-package ru.scratty.kinja.exception
-
-class ContextIsAlreadyInitializedException : RuntimeException("Context is already initialized")
