--- src/main/kotlin/ru/scratty/kinja/exception/MultipleObjectsWithTypeException.kt
@@ -1,6 +1,0 @@
-package ru.scratty.kinja.exception
-
-import kotlin.reflect.KClass
-
-class MultipleObjectsWithTypeException(clazz: KClass<*>) :
-    RuntimeException("Found multiple objects by class ${clazz.qualifiedName}")
