--- src/main/kotlin/ru/scratty/kinja/holder/Holder.kt
@@ -1,6 +1,0 @@
-package ru.scratty.kinja.holder
-
-interface Holder<T> {
-
-    fun get(): T
-}
