--- src/main/kotlin/ru/scratty/kinja/holder/SingleHolder.kt
@@ -1,6 +1,0 @@
-package ru.scratty.kinja.holder
-
-class SingleHolder<T>(private val obj: T) : Holder<T> {
-
-    override fun get(): T = obj
-}
