--- src/main/kotlin/ru/scratty/kinja/inject/InjectCandidate.kt
@@ -1,13 +1,0 @@
-package ru.scratty.kinja.inject
-
-import kotlin.reflect.KFunction
-
-data class InjectCandidate(
-    val name: String,
-    val func: KFunction<Any>,
-    val scope: InjectScope,
-) {
-
-    val params = func.parameters
-    val returnType = func.returnType
-}
