--- src/main/kotlin/ru/scratty/kinja/inject/InjectScope.kt
@@ -1,6 +1,0 @@
-package ru.scratty.kinja.inject
-
-enum class InjectScope {
-    SINGLE,
-    FACTORY,
-}
